Camera fixed +game end added + shader bug fixed
I investigated the distortion affecting the previously black-and-white silhouette on the screen and identified the shader as the cause. After replacing the shader, the issue was resolved.

Additionally, the camera issue on the organ map was fixed. A user interface was added for the endgame, and the falling-through-the-ground problem was resolved.